[INLONG-7089][Manager] Rename label table to stream_source_label
