+++ inlong-manager/manager-web/sql/apache_inlong_manager.sql
@@ -159,40 +159,6 @@ CREATE TABLE IF NOT EXISTS `inlong_cluster_node`
   DEFAULT CHARSET = utf8mb4 COMMENT ='Inlong cluster node table';
 
 -- ----------------------------
--- Table structure for inlong_label, label is used for filtering operations, such as which nodes are filtered when sending collection tasks
--- ----------------------------
-CREATE TABLE IF NOT EXISTS `inlong_label`
-(
-    `id`           int(11)      NOT NULL AUTO_INCREMENT COMMENT 'Incremental primary key',
-    `label_name`   varchar(128) NOT NULL COMMENT 'label name',
-    `ext_params`   mediumtext            DEFAULT NULL COMMENT 'Extended params, will be saved as JSON string',
-    `description`  varchar(256)          DEFAULT '' COMMENT 'Description of cluster tag',
-    `in_charges`   varchar(512) NOT NULL COMMENT 'Name of responsible person, separated by commas',
-    `is_deleted`   int(11)               DEFAULT '0' COMMENT 'Whether to delete, 0: not deleted, > 0: deleted',
-    `creator`      varchar(64)  NOT NULL COMMENT 'Creator name',
-    `modifier`     varchar(64)           DEFAULT NULL COMMENT 'Modifier name',
-    `create_time`  timestamp    NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT 'Create time',
-    `modify_time`  timestamp    NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT 'Modify time',
-    `version`      int(11)      NOT NULL DEFAULT '1' COMMENT 'Version number, which will be incremented by 1 after modification',
-    PRIMARY KEY (`id`),
-    UNIQUE KEY (`label_name`)
-) ENGINE = InnoDB
-  DEFAULT CHARSET = utf8mb4 COMMENT ='Inlong collect task label table';
-
--- ----------------------------
--- Table structure for inlong_label_node_relation
--- ----------------------------
-CREATE TABLE IF NOT EXISTS `inlong_label_node_relation`
-(
-    `id`            int(11)      NOT NULL AUTO_INCREMENT COMMENT 'Incremental primary key',
-    `label_id`      int(11)      NOT NULL COMMENT 'Id of the label id',
-    `node_id`       int(11)      NOT NULL COMMENT 'Id of the cluster node id.',
-    PRIMARY KEY (`id`),
-    UNIQUE KEY `unique_label_kv` (`label_id`, `node_id`)
-) ENGINE = InnoDB
-  DEFAULT CHARSET = utf8mb4 COMMENT ='Inlong collect task label table';
-
--- ----------------------------
 -- Table structure for inlong_consume
 -- ----------------------------
 CREATE TABLE IF NOT EXISTS `inlong_consume`
@@ -263,6 +229,40 @@ CREATE TABLE IF NOT EXISTS `stream_source_cmd_config`
     UNIQUE KEY `unique_source_cmd_config` (`task_id`, `bSend`, `specified_data_time`)
 ) ENGINE = InnoDB
   DEFAULT CHARSET = utf8;
+
+-- ----------------------------
+-- Table structure for stream_source_label, label is used for filtering operations, such as which nodes are filtered when sending collection tasks
+-- ----------------------------
+CREATE TABLE IF NOT EXISTS `stream_source_label`
+(
+    `id`           int(11)      NOT NULL AUTO_INCREMENT COMMENT 'Incremental primary key',
+    `label_name`   varchar(128) NOT NULL COMMENT 'label name',
+    `ext_params`   mediumtext            DEFAULT NULL COMMENT 'Extended params, will be saved as JSON string',
+    `description`  varchar(256)          DEFAULT '' COMMENT 'Description of cluster tag',
+    `in_charges`   varchar(512) NOT NULL COMMENT 'Name of responsible person, separated by commas',
+    `is_deleted`   int(11)               DEFAULT '0' COMMENT 'Whether to delete, 0: not deleted, > 0: deleted',
+    `creator`      varchar(64)  NOT NULL COMMENT 'Creator name',
+    `modifier`     varchar(64)           DEFAULT NULL COMMENT 'Modifier name',
+    `create_time`  timestamp    NOT NULL DEFAULT CURRENT_TIMESTAMP COMMENT 'Create time',
+    `modify_time`  timestamp    NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP COMMENT 'Modify time',
+    `version`      int(11)      NOT NULL DEFAULT '1' COMMENT 'Version number, which will be incremented by 1 after modification',
+    PRIMARY KEY (`id`),
+    UNIQUE KEY (`label_name`)
+    ) ENGINE = InnoDB
+    DEFAULT CHARSET = utf8mb4 COMMENT ='Inlong collect task label table';
+
+-- ----------------------------
+-- Table structure for stream_source_label_node_relation
+-- ----------------------------
+CREATE TABLE IF NOT EXISTS `stream_source_label_node_relation`
+(
+    `id`            int(11)      NOT NULL AUTO_INCREMENT COMMENT 'Incremental primary key',
+    `label_id`      int(11)      NOT NULL COMMENT 'Id of the label id',
+    `node_id`       int(11)      NOT NULL COMMENT 'Id of the cluster node id.',
+    PRIMARY KEY (`id`),
+    UNIQUE KEY `unique_label_kv` (`label_id`, `node_id`)
+    ) ENGINE = InnoDB
+    DEFAULT CHARSET = utf8mb4 COMMENT ='Inlong collect task label table';
 
 -- ----------------------------
 -- Table structure for inlong_stream
